pyproject.toml
[project]
name = "job_search"
version = "0.1.0"
description = "job_search using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.159.0,<1.0.0",
    "python-docx>=1.2.0",
    "python-jobspy>=1.1.13",
]

[project.scripts]
kickoff = "job_search.main:kickoff"
run_crew = "job_search.main:kickoff"
plot = "job_search.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
